src/requirements.txt
apache-airflow-providers-http
yfinance
numpy 
pandas
